Updated javascript_scenes.js and javascript_people.js
Health element scene begins with player going to healthcenter and finding it locked. Sahiya prompts her to obtain keys from three people. Every time she obtains one key, she goes to the healthcenter and is unable to enter. Sahiya tells her the importance of next key and where to find it. On obtaining all three keys, player is able to enter the healthcenter